+++ labs/memory_bound/data_packing/solution.h
@@ -13,9 +13,9 @@ constexpr int maxRandom = 100;
 // FIXME: this data structure can be reduced in size
 struct S {
   float d;
-  int l : 14;
-  short i : 7;
-  short s : 7;
+  unsigned int l : 14;
+  unsigned int i : 7;
+  unsigned short s : 7;
   bool b : 1;
 
   bool operator<(const S &s) const { return this->i < s.i; }
